Bugfix: Free variable check
There was a weird bug where PI bindings with rationals in their body
could sometimes cause aborts. The condition seemed to be that the PI
binding needed to have an unused parameter and the root cause was that
the `_free_in` function didn't handle rationals well.

Added a test to prevent regressions.