Bugfix: rational printing
Our code for printing out rationals was a bit odd. For positive
rationals is would print them twice, so the value `3` would be printed
as `33`, and `4/5` would be `4/54/5`.

For negative rationals it would print the number in standard format,
followed by the number in ~-negation format, so `-2` would be printed as
`-2(~ 2)`.

The core cause was that we had code that just printed rationals, and
code that case-split on the sign  and did fancy stuff like ~, and **we
ran both pieces of code**. It also looks like we managed to leak some
memory in the process. Good thing we don't do much printing :).