Add support for checking streams instead of files (#14)
This commit adds a variant of `check_file()` that takes an
`std::istream` instead of a file name. This allows proof checking
without creating temporary files. There are lots of opportunities for
cleaning up the code. However, this commit aims to introduce the feature
without changing the original code too much. Cleanup will follow in a
future commit.

One minor additional change: It changes the Clang-Format file to not try
to not use the most common alignment of `&` and `*` because otherwise
our formatting for LFSC is different from the one used in CVC4.