Fix parsing of identifiers at head of application
Fixes #20. When parsing the identifier at the head of an application, we
try to match it against reserved words such as `compare`. When this
matching fails, we have a prefix of an identifier that is also a prefix
of a reserved word, e.g. if our identifier is `compfoo` we have `comp`,
and we need to parse the rest of the identifier. If it happened that our
identifier is just a prefix of a reserved word, e.g. `compar`, then we'd
try to parse the rest of the identifier using `prefix_id()`, which skips
whitespace before an identifier and we ended up parsing the next
identifier as the suffix of the current identifier. This commit fixes
the issue by adding an argument to `prefix_id()` that determines whether
to skip whitespaces before an identifier or not. While this fixes the
issue, I intend to refactor the parsing code to first parse the complete
identifier and then to try matching it against known identifiers. This
would simplify the code and could even end up being slightly faster.