Better output on test failure
Before, we ran LFSC on the test and if there was an error, printed out
the `stdout` member from the CompletedProcess object.

That member is acutally a byte array, so it didn't print very nicely.
For example, newlines printed as '\n', instead of as newlines.

Now we decode (as UTF-8) the bytes before printing them.

This produces nicer output.