type-checker allows all proper types as pgm args
LFSC side conditions have argument types and return types.

Originally our type checker required these to by symbols that were
declared as types, or as functions to types.

This PR implements the following rule: any proper type can be the
argument of a program argument. This is a good rule, because it
keeps the dependent typing out of the side-conditions, but allows for
all reasonable, introspectable data to be handled in SCs.

To allow for the kind of program's found in the theory of quantifiers,
we continue to allow functions to types to be given as the type of a
parameter, and to be matched agaist. We'd like to adopt a more
principled approach to these kinds of programs, but that is orthogonal
to this change.

Implementation:
   * A function `compute_kind`.
   * Allow program argument types, return types, and match scrutinee
   types for which `compute_kind` returns `statType`. (i.e. proper
   types).
   * Adds test cases for
      * th_quant.plf
      * programs that take (term Real)
   * Fixes a random signed/unsigned comparison warning (on my compiler)