Annotated lambdas (#37)
This PR adds annotated lambdas (using the # sigil).
An example is found in the provided test.

This is necessary because we don't actually have annotated lambdas already. The closest thing is %, but they're really more like local declares, and are checked like that (without having function type, using tail recursion, etc.).