Apply weak-head-reduction in match.
For pattern-matching in a functional language, you mush reduce a
scrutinee before matching. Otherwise you don't know what value you're
matching the head against.

Even in haskell, the lay language, they have to do WHR to the scrutinee.
(this is the source of *all* eagerness in haskell's core).

Added a test to illustrate the problem.