Use LAM values rather than PI values for #-form (#41)
* Use LAM values rather than PI values for #-form

Additionally, use different symbols for the abstraction variables in the
type and value associated with a #-form.

Additionally, force these lambda value for a #-form to be cloned when
necessary (e.g. during weak-head-reduction, or during implicit
weak-head-reduction during stream-checking.

This fixes a class of bugs related to #-forms in which different
applications of the same form would interfere with each other and not be
properly applied during type-checking.

The root cause of these bugs is three-fold:
* LFSC does not apply weak-head-reduction to PI-values. I'm believe it
  should, but it suffices for us to use LAM-values in this instance.
* Using the same symbol for the abstraction variable in both the *value*
  and *type* of a #-form causes the abstraction-bindings created during
  type-checking to modify other terms in the system (those dependent) on
  the #-form's value in subtle and wrong ways.
* I don't fully understand the invariants that LFSC's no-clone WHR
  optimization relies on, but those invariants seem to be invalidated by
  the LAM-values we're creating here, so I've disabled that optimization
  for them.

* Increase testing stack size slightly

We were already pretty close to the max on one of the tests. This puts
us over.

* Output on test failure in CI

* Use pre-python3.6 in run_test.py

* clang-format