Macros in code expressions.
Previously, one could not use macros in code expressions.

An example:

(program fun ((t term) (s term)) term
  (and t s))

where `and` is a macro.

There wasn't a very good reason for this.
The only issue is that our type-checking machinery for code
assumed that all application heads were (a) programs or (b) declared
(and thus unbound) symbols.

A macro is just a bound symbol, which we do have type information stored
for, so this PR modifies the procedure for type-checking application
heads to look up that type information.