Evaluate code apps without weak-head-reduction
Previously, we weak-head-reduced code expressions *before* evaluating
them. This is non-standard, and I don't see why it should be done.

It has some specific negative effects. Namely, it embeds unevaluated
code expressions in the referent of symbols. Since code expressions and
base terms are not distinguished, this makes it hard to figure out
whether you actually want to "run" symbol referents in the future.
You should not do so, for base terms, but you should do so for code.
Even if you can figure it out, you might end up running the same code
multiple times, if it gets embedded in multiple places during
weak-head-reduction.

After this commit, we do the obvious thing: we evaluate code
applications! We *do not* weak-head reduce them...