Support older versions of CMake
`PREPEND` for lists was only introduced in CMake 3.15, but we require
cvc5 to build with CMake 3.9. This commit replaces the `PREPEND` by an
`APPEND`, adds an older version of Ubuntu as a CI test to catch these
issues in the future, and raises the minimum CMake version to 3.9 to
match cvc5's version requirements.

This change is similar to what we did in cvc5:
https://github.com/cvc5/cvc5/pull/7003